Experimenting with stored diags functionality, but not useful.
It's not useful because the stored diags remain the same after the
first parse of the translation unit (during the constructor of a
Session object). So when you introduce new errors in the cpp file
the diags won't get picked up for some reason. Not sure how to fix
this.