Fix crash when modifying a header in the preamble of a loaded session.
When a code complete request is started, the whole AST unit is reparsed.
This is really fast when no changes have been made in the preamble, so there is
no harm done. More importantly, this fixes a crash that occurs when a session
for an implementation file has been loaded, and then a header file in the
preamble of the implementation file is modified (for instance, new method or
attribute or whatever is added to a class in a header file). After the
modification, the `clang::ASTUnit` would crash for some reason when a new
code completion request is run. When reparsing the `clang::ASTUnit` just
_before_ the completion request, everything seems to be fine. Thus, when no
changes were made to the preamble, nothing is observably changed, but when a
change is made in the preamble, no code completions will occur until the
preamble is reparsed.